# target/classes/openapi/api-out.yaml
openapi: 3.0.1
info:
  title: ValkyrAI CORE API
  description: The API for accessing ValkyrAI services
  termsOfService: https://valkyrlabs.com/tos/
  contact: {email: info@valkyrlabs.com}
  license: {name: AGPL 3.0, url: 'https://www.gnu.org/licenses/agpl-3.0.html'}
  version: 0.9.25
...
servers:
- {url: 'http://localhost:8080/v1', description: local server (uses live data)}
- {url: 'https://valkyrlabs.com/v1', description: production server}
...
paths:
...
  /Logout:
    get:
      tags: [Logout]
      summary: Retrieve a list of Logouts
      description: Retrieves a list of Logouts.
      x-spring-cacheable: true
      operationId: getLogoutList
      responses:
        '200':
          description: Successfully retrieved the list of Logouts.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Logout'}
        '404': {description: No Logouts were found.}
        '500': {description: Internal server error.}
    post:
      tags: [Logout]
      summary: Create a new Logout
      description: Creates a new Logout.
      operationId: postLogout
      requestBody:
        description: Logout details.
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Logout'}
      responses:
        '201':
          description: Logout created successfully.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Logout'}
        '400': {description: Invalid input.}
        '404': {description: Logout not found.}
        '500': {description: Internal server error.}
  /Logout/{id}:
    get:
      tags: [Logout]
      summary: Retrieve a single Logout
      description: Retrieves a single Logout for a specific uid.
      x-spring-cacheable: true
      operationId: getLogout
      parameters:
      - name: id
        schema: {type: string, format: uuid}
        in: path
        required: true
        description: Unique identifier for the Logout.
      responses:
        '200':
          description: Successfully retrieved the Logout.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Logout'}
        '404': {description: Logout not found.}
        '500': {description: Internal server error.}
    put:
      tags: [Logout]
      summary: Update an existing Logout
      description: Updates an existing Logout.
      operationId: updateLogout
      parameters:
      - name: id
        schema: {type: string, format: uuid}
        in: path
        required: true
        description: Unique identifier for the Logout.
      requestBody:
        description: Updated Logout details.
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Logout'}
      responses:
        '200':
          description: Logout updated successfully.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Logout'}
        '400': {description: Invalid input.}
        '404': {description: Logout not found.}
        '500': {description: Internal server error.}
    delete:
      tags: [Logout]
      summary: Delete a Logout.
      description: Deletes a specific Logout.
      operationId: deleteLogout
      parameters:
      - name: id
        schema: {type: string, format: uuid}
        in: path
        required: true
        description: Unique identifier for the Logout.
      responses:
        '204': {description: Logout deleted successfully.}
        '404': {description: Logout not found.}
        '500': {description: Internal server error.}
...
components:
...
  schemas:
...
    Logout:
      type: object
      description: Logout functionality
      properties:
        description: {type: string}
        id:
          description: Unique identifier for object in the system
          type: string
          x-field-extra-annotation: "@Id  \n        @GeneratedValue(generator = \"\
            UUID\")\n"
          format: uuid
          example: 8ec508fe-99a6-4a98-b47e-e28d057b6284
        ownerId: {description: UUID of owner of the object in the system, type: string,
          x-field-extra-annotation: '        @AuditingField(fieldType = AuditingField.FieldType.CREATED_BY,
            enabled = true)', format: uuid, example: 29cfd419-c81a-47ae-8600-79ac5d8cc6c7}
        createdDate: {description: Date of object creation, type: string, x-field-extra-annotation: '        @AuditingField(fieldType
            = AuditingField.FieldType.CREATED_DATE, enabled = true)', format: date-time,
          example: '2025-01-03T16:21:41.596-0800'}
        keyHash: {description: 'Data, including hash of the key(s) used to encrypt
            this record.', type: string}
        lastAccessedById: {description: Last user to access object, type: string,
          x-field-extra-annotation: '        @AuditingField(fieldType = AuditingField.FieldType.LAST_ACCESSED_BY,
            enabled = true)', format: uuid, example: 93850cc7-673e-48ea-a594-385bde8e9fca}
        lastAccessedDate: {description: Timestamp of last access of object, type: string,
          x-field-extra-annotation: '        @AuditingField(fieldType = AuditingField.FieldType.LAST_ACCESSED_DATE,
            enabled = true)', format: date-time, example: '2025-01-03T16:21:41.596-0800'}
        lastModifiedById: {description: Unique identifier for user who last modifed
            the object in the system, type: string, x-field-extra-annotation: '        @AuditingField(fieldType
            = AuditingField.FieldType.LAST_MODIFIED_BY, enabled = true)', format: uuid,
          example: 60418982-a1ea-4832-8848-d37adf120363}
        lastModifiedDate: {description: Date of last object modification, type: string,
          x-field-extra-annotation: '    @AuditingField(fieldType = AuditingField.FieldType.LAST_MODIFIED_DATE,
            enabled = true)', format: date-time, example: '2025-01-03T16:21:41.596-0800'}
...
  securitySchemes:
    thorapi_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://{username}.valkyrlabs.com:{port}/oauth
          scopes: {'write:items': modify owned items, 'read:items': read owned items}